src/requirements.txt
fastapi
jinja2
uvicorn
python-dateutil
